BOL-124 refactor: 커서 관련 클래스 디렉토리 이동
