+++ adapter-in/web/src/main/kotlin/com/yapp/bol/group/member/MemberController.kt
@@ -11,8 +11,8 @@ import com.yapp.bol.group.member.dto.JoinGroupRequest
 import com.yapp.bol.group.member.dto.MemberResponse
 import com.yapp.bol.group.member.dto.ValidateMemberNameResponse
 import com.yapp.bol.group.member.dto.toResponse
-import com.yapp.bol.pagination.SimpleCursorResponse
-import com.yapp.bol.pagination.group.member.MemberCursorRequest
+import com.yapp.bol.pagination.cursor.SimpleCursorResponse
+import com.yapp.bol.pagination.cursor.member.MemberCursorRequest
 import org.springframework.security.access.prepost.PreAuthorize
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PathVariable
