+++ adapter-out/rdb/src/main/kotlin/com/yapp/bol/group/member/CustomMemberRepositoryImpl.kt
@@ -1,7 +1,7 @@
 package com.yapp.bol.group.member
 
 import com.querydsl.jpa.impl.JPAQueryFactory
-import com.yapp.bol.pagination.CursorRequest
+import com.yapp.bol.pagination.cursor.CursorRequest
 
 class CustomMemberRepositoryImpl(
     val queryFactory: JPAQueryFactory
