+++ domain/src/main/kotlin/com/yapp/bol/pagination/cursor/SimpleCursorRequest.kt
@@ -1,4 +1,4 @@
-package com.yapp.bol.pagination
+package com.yapp.bol.pagination.cursor
 
 data class SimpleCursorRequest<CURSOR>(
     override val size: Int,
