+++ domain/src/main/kotlin/com/yapp/bol/pagination/cursor/member/MemberCursorRequest.kt
@@ -1,7 +1,7 @@
-package com.yapp.bol.pagination.group.member
+package com.yapp.bol.pagination.cursor.member
 
 import com.yapp.bol.group.GroupId
-import com.yapp.bol.pagination.CursorRequest
+import com.yapp.bol.pagination.cursor.CursorRequest
 
 data class MemberCursorRequest(
     val groupId: GroupId,
