+++ domain/src/main/kotlin/com/yapp/bol/pagination/offset/PaginationCursor.kt
@@ -1,4 +1,4 @@
-package com.yapp.bol.pageable
+package com.yapp.bol.pagination.offset
 
 data class PaginationCursor<T>(
     val content: List<T>,
