+++ port-in/src/main/kotlin/com/yapp/bol/group/GroupService.kt
@@ -5,12 +5,13 @@ import com.yapp.bol.group.dto.CreateGroupDto
 import com.yapp.bol.group.dto.GroupMemberList
 import com.yapp.bol.group.dto.GroupWithMemberCount
 import com.yapp.bol.group.dto.JoinGroupDto
-import com.yapp.bol.pageable.PaginationCursor
+import com.yapp.bol.pagination.offset.PaginationCursor
 
 interface GroupService {
     fun createGroup(
         createGroupDto: CreateGroupDto
     ): GroupMemberList
+
     fun joinGroup(request: JoinGroupDto)
 
     fun searchGroup(
